Ubicacion Agregada a la Nota junto con mapa
Ahora el usuario puede agregar su ubicacion al crear una nota, ademas, puede ver en el mapa esta ubicacion tanto al crear como al editar la nota

+++ lib/dominio/agregados/VONota/VOUbicacionNota.dart
@@ -1,18 +1,18 @@
 class VOUbicacionNota {
-  int latitud;
-  int longitud;
+  double latitud;
+  double longitud;
 
   VOUbicacionNota(this.latitud, this.longitud);
 
-  static VOUbicacionNota crearUbicacionNota(int latitud, int longitud) {
+  static VOUbicacionNota crearUbicacionNota(double latitud, double longitud) {
     return VOUbicacionNota(latitud, longitud);
   }
 
-  int getLatitud() {
+  double getLatitud() {
     return latitud;
   }
 
-  int getLongitud() {
+  double getLongitud() {
     return longitud;
   }
 }
